Moved the device idea integrated BT serial module into the comm directory.
