+++ main/comm/BTspp.cpp
@@ -1,4 +1,4 @@
-#include "BTSender.h"
+#include "BTspp.h"
 #include "SetupNG.h"
 
 #include "comm/DataLink.h"
